`RuleMetaData` with lifetime data
Lifetime/expiration data is now part of `RuleMetaData`, in a similar
fashion to what was done to content settings in the past. The existing
code sets expiration with an empty lifetime delta. Brave doesn't use
upstream chromium's expiration mechanism, however this change preserves
in place the passing of the expiration data in these types.

Chromium changes:
https://chromium.googlesource.com/chromium/src/+/8e3f041de669e879c4234549091bf3e22ad30c4a

commit 8e3f041de669e879c4234549091bf3e22ad30c4a
Author: Chris Fredrickson <cfredric@chromium.org>
Date:   Mon Jun 19 13:05:35 2023 +0000

    Add lifetime field to RuleMetaData

    Tracking the lifetime explicitly in the RuleMetaData means that each
    ContentSettingPatternSource now has enough information to be
    "refreshed"/renewed, if desired.

    Bug: 1450356