Add iron-dropdown to the build.
We use cr_searchable_drop_down in the wallet's Add dialog. That control
is only available in ChromeOS upstream so we've been patching upstream
build to include it for Desktop. Now with the bundle_js migration
upstream we also need to add iron_dropdown to our build because
cr_searchable_drop_down depends on it.