[Android] New c-tor signature for TabbedRootUiCoordinator
Chromium change:
https://github.com/chromium/chromium/commit/8fe1cbc3480be673ef4fe7fd3f71a13b97a6ef0c

Incognito Reauth: Fix reauth screen not restored when activity is killed
Under memory pressure (emulated by the "Don't keep activities" setting), IncognitoReauthController gets killed after starting the reauthentication flow. When restored, savedInstanceState is checked for incognitoReauthPending when re-initializing the IncognitoReauthController, to determine whether to resume the reauthentication flow.

The problem lies when the SavedInstanceState is reset in
ChromeTabbedActivity::onStartWithNative which sets it to null before it
is checked for the incognitoReauthPending key in
RootUiCoordinator::onFinishNativeInitialization, making
mIsIncognitoReauthPendingOnRestore in IncognitoReauthController always
return false and the reauth screen not getting restored when the
activity is killed.

This CL saves the incognitoReauthPending key value from
SavedInstanceState in the the constructor of RootUiCoordinator before
it is reset.

(cherry picked from commit 318d0b3)

Demo: https://drive.google.com/file/d/16nBDAFA3bIk065EsGGmynMLTY_zU7Uop/view?usp=sharing
Bug: 1365633