Remove redundant `NavigationManagerImpl` delegate setting
Upstream has changed `NavigationManagerImpl`'s constructor to expect the
delegate to be passed during construction. For some unknown reason, the
iOS code in Brave was setting the delegate for the instance in the
`web::WebState` instance by itself, even though this seems to be of no
advantage whatsoever.

This change removes the unecessary operation, as it cannot be supported
anymore.

Chromium change:
https://chromium.googlesource.com/chromium/src/+/c624e1a7f9608dce4ffe5fde9e3afe883a88d7a2

commit c624e1a7f9608dce4ffe5fde9e3afe883a88d7a2
Author: Sylvain Defresne <sdefresne@chromium.org>
Date:   Mon Jun 26 09:24:36 2023 +0000

    [ios] Make BrowserState and Delegate const in NavigationManagerImpl

    Change NavigationManagerImpl constructor to take the pointer to the
    BrowserState and Delegate so that the member variable can be marked
    as const.

    Remove the setter and cleanup client code and tests.

    Bug: none