don't use TestFiltersProvider as a ResourceProvider
Originally I intended this commit to remove `UseSourceProviderForTest`
and `UseCustomSourceProviderForTest`, but I underestimated the number of
places they were copy-pasted outside of `AdBlockServiceTest`. The
replacement methods match real-world behavior more closely, but require
a bit more boilerplate that I didn't want to continue duplicating
everywhere. I'll return to that later.

I've also removed `HTTPSEverywhereServiceTest.PreferAdblockRedirect` in
this commit, since it was the only test outside of `AdBlockServiceTest`
that actually used adblock resources. The HTTPS Everywhere component is
no longer used and will be removed soon, so this shouldn't be too
controversial.