Working around `static_assert` cancellation
There's an override to cancel the use of `static_assert` in the
shadowing file changed in this commit. This has caused a build failure
on an upstream file that necessitates the static assert to be fully
evaluated.

The existing `static_assert` substitution shouldn't exist in the first
place, and the author who introduced this override has created an issue
to correct this use. In the meantime, this change preempts the inclusion
of the header with the build failure.

Chromium change:
https://chromium.googlesource.com/chromium/src/+/9b396d355c49fc980f4dbcddc745d77a63a0caf1

commit 9b396d355c49fc980f4dbcddc745d77a63a0caf1
Author: Adam Rice <ricea@chromium.org>
Date:   Mon Sep 25 07:13:42 2023 +0000

    Optimize net::HttpUtil::IsLWS() for Android

    Clang is unable to effectively optimize IsLWS() on Android because it
    is compiled with -Oz. This results in large and inefficient code.

    Force inlining and manually optimise the function to get good
    performance on Android.